Длительность фильма форматируется в шаблоне
